Move `TransferSchema` and make `tokenAddress` required
